+++ src/webview/model-editor/components/index.ts
@@ -1,0 +1,10 @@
+export { InfoTip } from './InfoTip';
+export { TabBar } from './TabBar';
+export { DirtyField } from './DirtyField';
+export { TextField } from './TextField';
+export { TextListField } from './TextListField';
+export { InitNextFields } from './InitNextFields';
+export { TextArea } from './TextArea';
+export { ModelNameEditor } from './ModelNameEditor';
+export { NumberField } from './NumberField';
+export { placeholderForKind } from './helpers';
